fixing for heroku

+++ app/models/stock_yank.rb
@@ -25,8 +25,10 @@ class StockYank < ActiveRecord::Base
   end
 
   def sort_data_and_save
-    StockYank.delete_all
-    StockYank.connection.execute 'delete from sqlite_sequence where name = "stock_yanks"'
+    if ! StockYank.find(:first)
+      StockYank.delete_all
+      StockYank.connection.execute 'delete from sqlite_sequence where name = "stock_yanks"'
+    end
     workstring = (yank_data/"tbody").to_html
     stock_array = Array.new
     for i in (0..74)
